Fix "not a supported format" error for Linux
On launch the argv list was examined to see if it contains the
executable as the first entry.  This was done because windows may or
may not contain it.  The test to see if the first element of argv was
the executable failed when called from a symlink.

This fix uses the QCommandLineParser to handle this possibility more
robustly.